+++ R/get_info_oak.R
@@ -4,7 +4,7 @@
 #' This function retrieves information about a specific oak tree based on its ID
 #' from the provided oak tree database.
 #'
-#' @param dbchene Database containing information about oak trees.
+#' @param dboak Database containing information about oak trees.
 #' @param theidoak ID of the oak tree for which information is to be retrieved.
 #'
 #' @return A list containing information about the oak tree, including its type and
@@ -22,14 +22,14 @@
 #'  )
 #' chene <- DBI::dbReadTable(conn, name = "chenes")
 #'
-#' get_info_chene(dbchene = chene, theidoak = "162")
+#' get_info_oak(dboak = chene, theidoak = "162")
 #'
-get_info_chene <- function(dbchene, theidoak) {
-  check_param(dbchene, "data.frame")
+get_info_oak <- function(dboak, theidoak) {
+  check_param(dboak, "data.frame")
   check_param(theidoak, "character")
-  check_names_dataframe(c("idoak", "type", "planting_date"), dbchene)
+  check_names_dataframe(c("idoak", "type", "planting_date"), dboak)
 
-  oak <- dbchene |>
+  oak <- dboak |>
     filter(idoak == theidoak)
 
   return(list(