+++ NEWS.md
@@ -1,9 +1,15 @@
 # truffles (development version)
 
+- Rename `update_db_truffe()` -> `update_db_truffle()`
+- Rename `write_db_new_truffe()` -> `write_db_new_truffle()`
+- Rename `get_info_chene()` -> `get_info_oak()`
+- Rename `get_info_chene_last_truffe()` -> `get_info_oak_last_truffle()`
+- Rename `get_info_chene_truffe()` -> `get_info_oak_truffle()`
+
 # truffles 0.0.2
 
 - Using flexbox in modals
-- New function get_info_reensemence() to get information about the latest reseeding date
+- New function `get_info_reensemence()` to get information about the latest reseeding date
 - Add User Doc / Dev Doc in pkgdown 
 
 # truffles 0.0.1
@@ -15,4 +21,4 @@
 - New function get_info_chene_truffe() to get information about truffles associated with an oak tree
 - New function update_db_truffe() to update a row in the "truffe" table of a database.
 - New function weight_truffles_by() to calculate total weight of truffles grouped by specified columns
-- New function write_db_new_truffe() to write new truffle data to the database
+- New function write_db_new_truffle() to write new truffle data to the database